requirements.txt
requests
beautifulsoup4
python-dotenv
mysql-connector-python
